Separate reusable components (not connected to the application store) from other components
