+++ src/components/Playlists/MosaicMode/styles.css
@@ -28,3 +28,7 @@
     overflow: hidden;
     width: 100%;
 }
+
+.divider {
+    margin: 5px 0px 5px 0px;
+}
