+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './utils/translations/i18n';
-import App from './components/App';
+import App from './containers/App';
 
 ReactDOM.render(
     <React.StrictMode>
